module-7-data-governance-optional/requirements.txt
scikit-learn
pandas
numpy
pyyaml
dvc[s3]
